Small fixes
Added port to incoming messages(on clients both the ip and port are now
shown)
When a new client is connected, the listbox will clear before
repopulating it.
When sending a message the textbox will clear.
The program will no longer hang when connecting to a server if the
server's info are incorrect.